Move _get_response_factory to pyramid.response
