+++ pyramid/tests/test_response.py
@@ -8,7 +8,7 @@ class TestResponse(unittest.TestCase):
     def _getTargetClass(self):
         from pyramid.response import Response
         return Response
-        
+
     def test_implements_IResponse(self):
         from pyramid.interfaces import IResponse
         cls = self._getTargetClass()
@@ -119,7 +119,7 @@ class Test_patch_mimetypes(unittest.TestCase):
         result = self._callFUT(module)
         self.assertEqual(result, True)
         self.assertEqual(module.initted, True)
-        
+
     def test_missing_init(self):
         class DummyMimetypes(object):
             pass
@@ -174,6 +174,17 @@ class TestResponseAdapter(unittest.TestCase):
         self.assertEqual(dummy_venusian.attached,
                          [(foo, dec.register, 'pyramid')])
 
+
+class TestGetResponseFactory(unittest.TestCase):
+    def test_get_factory(self):
+        from pyramid.registry import Registry
+        from pyramid.response import Response, _get_response_factory
+
+        registry = Registry()
+        response = _get_response_factory(registry)(None)
+        self.assertTrue(isinstance(response, Response))
+
+
 class Dummy(object):
     pass
 
@@ -190,5 +201,3 @@ class DummyVenusian(object):
 
     def attach(self, wrapped, fn, category=None):
         self.attached.append((wrapped, fn, category))
-
-
