+++ pyramid/util.py
@@ -555,15 +555,3 @@ def action_method(wrapped):
         functools.update_wrapper(wrapper, wrapped)
     wrapper.__docobj__ = wrapped
     return wrapper
-
-
-def _get_response_factory(registry):
-    """ Obtain a :class: `pyramid.response.Response` using the
-    `pyramid.interfaces.IResponseFactory`.
-    """
-    response_factory = registry.queryUtility(
-        IResponseFactory,
-        default=lambda r: Response()
-    )
-
-    return response_factory
